+++ rotkehlchen/tests/api/test_assets.py
@@ -696,7 +696,7 @@ def test_search_assets(rotkehlchen_api_server):
         },
     )
     result = assert_proper_sync_response_with_result(response)
-    assert len(result) == 5
+    assert len(result) == 6
     assert any(entry['name'] == 'Ethereum' for entry in result)
     for entry in result:
         assert entry['symbol'] == 'ETH'
@@ -724,7 +724,7 @@ def test_search_assets(rotkehlchen_api_server):
         },
     )
     result = assert_proper_sync_response_with_result(response)
-    assert len(result) == 5
+    assert len(result) == 6
     assert any(entry['name'] == 'Ethereum' for entry in result)
     for entry in result:
         assert entry['symbol'] == 'ETH'
@@ -737,6 +737,8 @@ def test_search_assets(rotkehlchen_api_server):
             assert entry['evm_chain'] == 'base'
         elif entry['identifier'].startswith('eip155:137'):
             assert entry['evm_chain'] == 'polygon_pos'
+        elif entry['identifier'].startswith('eip155:250'):
+            assert entry['evm_chain'] == 'fantom'
         else:
             assert 'evm_chain' not in entry
     assert_asset_result_order(data=result, is_ascending=True, order_field='name')
@@ -783,6 +785,7 @@ def test_search_assets(rotkehlchen_api_server):
         'base',
         'scroll',
         'gnosis',
+        'fantom',
     }
 
     # check that using evm_chain filter works.