+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :bug:`-` Deposit events in Aave will now show in the correct order.
 * :feature:`3971` Show the total collateral ratio in the Liquity Trove section.
 * :feature:`2323` Add support for Huobi exchange.
 * :feature:`-` Curve vote escrow locking and withdrawal events will now be properly decoded on ethereum mainnet.